refactor: infrastructure Layer 도입
