+++ src/main/java/kitchenpos/menu/domain/MenuDao.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.repository;
+package kitchenpos.menu.domain;
 
 import java.util.List;
 import java.util.Optional;
