+++ src/main/java/kitchenpos/menu/infrastructure/JpaMenuProductDao.java
@@ -1,8 +1,9 @@
-package kitchenpos.menu.repository;
+package kitchenpos.menu.infrastructure;
 
 import java.util.List;
 import java.util.Optional;
 import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProductDao;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Repository;
