+++ src/main/java/kitchenpos/menu/infrastructure/JpaMenuProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.repository;
+package kitchenpos.menu.infrastructure;
 
 import java.util.List;
 import kitchenpos.menu.domain.MenuProduct;
