+++ src/main/java/kitchenpos/menugroup/infrastructure/JpaMenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menugroup.repository;
+package kitchenpos.menugroup.infrastructure;
 
 import kitchenpos.menugroup.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
