+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -3,15 +3,15 @@ package kitchenpos.order.application;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import kitchenpos.menu.repository.MenuDao;
+import kitchenpos.menu.domain.MenuDao;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.dto.OrderCreateRequestDto;
 import kitchenpos.order.dto.OrderCreateResponseDto;
 import kitchenpos.order.dto.OrderLineItemDto;
-import kitchenpos.order.repository.OrderDao;
-import kitchenpos.table.repository.OrderTableDao;
+import kitchenpos.order.domain.OrderDao;
+import kitchenpos.table.domain.OrderTableDao;
 import kitchenpos.table.domain.OrderTable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
