+++ src/main/java/kitchenpos/order/domain/OrderDao.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.repository;
+package kitchenpos.order.domain;
 
 import java.util.List;
 import java.util.Optional;
