+++ src/main/java/kitchenpos/product/infrastructure/JpaProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.repository;
+package kitchenpos.product.infrastructure;
 
 import kitchenpos.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
