+++ src/main/java/kitchenpos/table/application/TableGroupService.java
@@ -3,9 +3,9 @@ package kitchenpos.table.application;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.order.repository.OrderDao;
-import kitchenpos.table.repository.OrderTableDao;
-import kitchenpos.table.repository.TableGroupDao;
+import kitchenpos.order.domain.OrderDao;
+import kitchenpos.table.domain.OrderTableDao;
+import kitchenpos.table.domain.TableGroupDao;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
