+++ src/main/java/kitchenpos/table/infrastructure/JpaOrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.repository;
+package kitchenpos.table.infrastructure;
 
 import java.util.List;
 import kitchenpos.table.domain.OrderTable;
