+++ src/main/java/kitchenpos/table/infrastructure/JpaTableGroupDao.java
@@ -1,8 +1,9 @@
-package kitchenpos.table.repository;
+package kitchenpos.table.infrastructure;
 
 import java.util.List;
 import java.util.Optional;
 import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.domain.TableGroupDao;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Repository;
