+++ src/test/java/kitchenpos/dao/OrderTableDaoTest.java
@@ -3,10 +3,10 @@ package kitchenpos.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Arrays;
-import kitchenpos.table.repository.OrderTableDao;
+import kitchenpos.table.domain.OrderTableDao;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
-import kitchenpos.table.repository.TableGroupDao;
+import kitchenpos.table.domain.TableGroupDao;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
