+++ src/test/java/kitchenpos/dao/TableGroupDaoTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDateTime;
 import kitchenpos.table.domain.TableGroup;
-import kitchenpos.table.repository.TableGroupDao;
+import kitchenpos.table.domain.TableGroupDao;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
