+++ src/test/java/kitchenpos/integration/FixtureMaker.java
@@ -4,14 +4,14 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.List;
-import kitchenpos.menu.repository.MenuDao;
-import kitchenpos.menugroup.repository.MenuGroupDao;
-import kitchenpos.menu.repository.MenuProductDao;
-import kitchenpos.order.repository.OrderDao;
-import kitchenpos.order.repository.OrderLineItemDao;
-import kitchenpos.table.repository.OrderTableDao;
-import kitchenpos.product.repository.ProductDao;
-import kitchenpos.table.repository.TableGroupDao;
+import kitchenpos.menu.domain.MenuDao;
+import kitchenpos.menugroup.domain.MenuGroupDao;
+import kitchenpos.menu.domain.MenuProductDao;
+import kitchenpos.order.domain.OrderDao;
+import kitchenpos.order.domain.OrderLineItemDao;
+import kitchenpos.table.domain.OrderTableDao;
+import kitchenpos.product.domain.ProductDao;
+import kitchenpos.table.domain.TableGroupDao;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
